refactor: moved getLastSessionEndBlock helper function to utils/block
